Error checks and general upgrades added.
A couple of issues had been brought to my attention.  One pertained to the lack of NULL checks.  Another pointed out that I should really update the function signatures so that they properly reflected the access types of the char pointers being passed into the functions.

I revamped the primary data structure, struct map_t, so that no longer served as a node in a link list.  A new data structure, struct map_link_t, was created as a replacement.  struct map_t now acts as an anchor point for the list.  It also houses error code and error line information for errors that occur in the function calls. I didn't want to write error messages to stderr directly or anything like that, so the approach is to call a function and then to check the error_code value upon return.  If the error_code entry is zero, the function should have succeeded.  Please refer to the test program map_stuff.c to see how the error checks have been instrumented into the calling code.

I added a Windows batch file map_comp.bat as I no longer have a working Mac to use to compile that version of the library.